Move RenderThread template to a separate header.
