+++ src/apps/snacman/snacman/CMakeLists.txt
@@ -12,6 +12,7 @@ set(${TARGET_NAME}_HEADERS
     Input.h
     Logging.h
     LoopSettings.h
+    RenderThread.h
     SparseSet.h
     Timing.h
